requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.3
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.7.15=haa95532_0
certifi=2025.8.3=pypi_0
charset-normalizer=3.4.3=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.3.3=pypi_0
cycler=0.12.1=pypi_0
expat=2.7.1=h8ddb27b_0
filelock=3.19.1=pypi_0
flatbuffers=25.2.10=pypi_0
fonttools=4.59.1=pypi_0
fsspec=2025.7.0=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
jinja2=3.1.6=pypi_0
kiwisolver=1.4.9=pypi_0
libffi=3.4.4=hd77b12b_1
markupsafe=3.0.2=pypi_0
matplotlib=3.10.5=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.5=pypi_0
numpy=2.2.6=pypi_0
onnxruntime-gpu=1.22.0=pypi_0
opencv-python=4.12.0.88=pypi_0
openssl=3.0.17=h35632f6_0
packaging=25.0=pypi_0
pandas=2.3.2=pypi_0
pillow=11.3.0=pypi_0
pip=25.1=pyhc872135_2
protobuf=6.32.0=pypi_0
psutil=7.0.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyparsing=3.2.3=pypi_0
pyreadline3=3.5.4=pypi_0
python=3.12.11=h716150d_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
requests=2.32.5=pypi_0
scipy=1.16.1=pypi_0
setuptools=78.1.1=py312haa95532_0
six=1.17.0=pypi_0
sqlite=3.50.2=hda9a48d_1
sympy=1.14.0=pypi_0
tk=8.6.15=hf199647_0
torch=2.8.0=pypi_0
torchvision=0.23.0=pypi_0
tqdm=4.67.1=pypi_0
typing-extensions=4.14.1=pypi_0
tzdata=2025.2=pypi_0
ucrt=10.0.22621.0=haa95532_0
ultralytics=8.3.184=pypi_0
ultralytics-thop=2.0.16=pypi_0
urllib3=2.5.0=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
wheel=0.45.1=py312haa95532_0
xz=5.6.4=h4754444_1
zlib=1.2.13=h8cc25b3_1
